Move `Environment` to server package.
