+++ server/src/main/java/org/spine3/server/Environment.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.spine3.util;
+package org.spine3.server;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Optional;
