+++ server/src/main/java/org/spine3/server/commandbus/CommandBus.java
@@ -36,7 +36,7 @@ import org.spine3.server.bus.Bus;
 import org.spine3.server.commandstore.CommandStore;
 import org.spine3.server.failure.FailureBus;
 import org.spine3.type.CommandClass;
-import org.spine3.util.Environment;
+import org.spine3.server.Environment;
 
 import javax.annotation.Nullable;
 import java.util.List;
