+++ server/src/test/java/org/spine3/server/EnvironmentShould.java
@@ -18,13 +18,14 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.spine3.util;
+package org.spine3.server;
 
 import org.junit.After;
 import org.junit.AfterClass;
 import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
+import org.spine3.server.Environment;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNull;
