+++ server/src/test/java/org/spine3/server/storage/StorageFactorySwitchShould.java
@@ -27,7 +27,7 @@ import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.spine3.server.storage.memory.InMemoryStorageFactory;
-import org.spine3.util.Environment;
+import org.spine3.server.Environment;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
